submit

+++ srcs/ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: macbook_air <macbook_air@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/22 15:47:45 by macbook_air       #+#    #+#             */
-/*   Updated: 2021/11/26 06:19:17 by macbook_air      ###   ########.fr       */
+/*   Updated: 2021/11/28 23:39:23 by macbook_air      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	char			*str;
-	unsigned int	i;
+	char	*str;
+	size_t	i;
 
 	if (!s || !f)
 		return (NULL);
